refactor: :recycle: refactor webhook responses with WebhookResponseFactory
