+++ apps/taxes/src/pages/api/webhooks/order-created.ts
@@ -12,6 +12,7 @@ import { createLogger } from "../../../lib/logger";
 import { getActiveTaxProvider } from "../../../modules/taxes/active-tax-provider";
 import { createClient } from "../../../lib/graphql";
 import { Client } from "urql";
+import { WebhookResponseFactory } from "../../../modules/app/webhook-response-factory";
 
 export const config = {
   api: {
@@ -65,6 +66,7 @@ export default orderCreatedAsyncWebhook.createHandler(async (req, res, ctx) => {
   const logger = createLogger({ event: ctx.event });
   const { payload, authData } = ctx;
   const { saleorApiUrl, token } = authData;
+  const webhookResponse = new WebhookResponseFactory(res);
 
   logger.info({ payload }, "Handler called with payload");
 
@@ -76,7 +78,7 @@ export default orderCreatedAsyncWebhook.createHandler(async (req, res, ctx) => {
     if (!activeTaxProvider.ok) {
       logger.error({ error: activeTaxProvider.error }, "getActiveTaxProvider error");
       logger.info("Returning no data");
-      return res.status(200).json({ success: false });
+      return webhookResponse.failureNoRetry(activeTaxProvider.error);
     }
 
     logger.info({ activeTaxProvider }, "Fetched activeTaxProvider");
@@ -101,11 +103,9 @@ export default orderCreatedAsyncWebhook.createHandler(async (req, res, ctx) => {
     await updateOrderMetadataWithExternalId(client, payload.order.id, createdOrder.id);
     logger.info("Updated order metadata with externalId");
 
-    return res.status(200).json({ success: true });
+    return webhookResponse.success();
   } catch (error) {
     logger.error({ error }, "Error while creating order in tax provider");
-    return res
-      .status(400)
-      .json({ success: false, message: "Error while creating order in tax provider" });
+    return webhookResponse.failureRetry("Error while creating order in tax provider");
   }
 });
